advisor refactor
